adjust level

+++ serieIV/quadratic.c
@@ -1,19 +1,21 @@
+// Calculate the roots of a quadratic equation.
+
 #include <stdio.h>
 #include <math.h>
 
 int main()
 {
-    double a, b, c;
-    double x1, x2;
-    double discriminant;
+    int a, b, c;
+    float x1, x2;
+    int discriminant;
 
 // Read in the coefficients of the equation
     printf("Enter the value of a: ");
-    scanf("%lf", &a);
+    scanf("%d", &a);
     printf("Enter the value of b: ");
-    scanf("%lf", &b);
+    scanf("%d", &b);
     printf("Enter the value of c: ");
-    scanf("%lf", &c);
+    scanf("%d", &c);
 
 // Calculate the discriminant
     discriminant = b * b - 4 * a * c;
@@ -23,15 +25,14 @@ int main()
         printf("There are no real solutions.\n");
     } else if (discriminant == 0) {
         // There is one real solution
-        x1 = -b / (2 * a);
+        x1 = -b / (float)(2 * a);
         printf("There is one real solution: x = %lf\n", x1);
     } else {
         // There are two real solutions
-        x1 = (-b + sqrt(discriminant)) / (2 * a);
-        x2 = (-b - sqrt(discriminant)) / (2 * a);
+        x1 = (-b + sqrt(discriminant)) / (float)(2 * a);
+        x2 = (-b - sqrt(discriminant)) / (float)(2 * a);
         printf("There are two real solutions: x1 = %lf, x2 = %lf\n", x1, x2);
     }
 
     return 0;
-
 }
